feat(ocpp): add OCPP 2.1 support
Add a selectable OCPP 2.1 version alongside 1.6 and 2.0.1. 2.1 is a strict
superset of 2.0.1 and the simulator behaves identically for the core Calls,
so the charging-station and CSMS bindings are shared rather than duplicated.

ferrowl-ocpp (wire layer):
- New `v2_1` feature (added to default) wiring `rust-ocpp/v2_1`.
- New action/v2_1.rs: a 90-row `define_ocpp_version!` table over
  `rust_ocpp::v2_1`, advertising the `ocpp2.1` subprotocol, with the full
  cs/csms partition and connector-scope tags. The 26 new 2.1 actions
  (tariffs, DER control, battery swap, periodic event streams, dynamic
  schedules) are categorised by message direction. `NotifyPeriodicEventStream`
  is omitted: it is a one-way stream datagram with no Request/Response pair.
- Export `Action21`/`Response21`/`V2_1` and the `v2_1` module.

ferrowl (TUI):
- Extract the 2.0.1 inbound handler and the ClientVersion/ServerVersion
  bodies into `macro_rules!` macros (client/server `v2_common.rs`) and
  instantiate them for both V2_0_1 and V2_1. The four typed client responses
  (Get/SetVariables, Reset, SetChargingProfile) are field-identical across the
  two versions, so only the rust_ocpp module and the marker/enum types vary.
  The macros take plain idents (`:ident` can precede `::`, `:path` cannot).
- 2.1 reuses 2.0.1's shared state types and action specs; the new
  client/v2_1 and server/v2_1 modules are thin macro invocations.
- Add `OcppVersion::V2_1` (config, Display, setup-dialog selector), dispatch
  arms in the client/server view builders, and two demo tabs (port 9002).

Add tests/ws_loopback_v21.rs: a real-websocket loopback proving the `ocpp2.1`
subprotocol is negotiated end-to-end with a bidirectional Call round-trip.
