+++ ferrowl/src/module/ocpp/server/v2_0_1/handler.rs
@@ -1,146 +1,35 @@
-//! OCPP 2.0.1 inbound (CS→CSMS) handler for the CSMS role. Default-derived Accepted/empty replies,
-//! except BootNotification (Accepted + currentTime + interval) and Heartbeat (currentTime).
-//! Authorize and a transaction-starting TransactionEvent (one carrying an idToken) gate the tag
-//! against the RFID accept-lists: Authorize (no EVSE) against the CS list unioned with every
-//! connector list, TransactionEvent against its EVSE's list ∪ the CS list. Every Call + reply is
-//! forwarded to the view.
+//! OCPP 2.0.1 inbound (CS→CSMS) handler for the CSMS role. The handler body is shared with 2.1 via
+//! [`define_csms_handler!`](crate::module::ocpp::server::v2_common::define_csms_handler) and is
+//! instantiated here for `V2_0_1`. Default-derived Accepted/empty replies, except BootNotification,
+//! Heartbeat, and the RFID-gated Authorize / transaction-starting TransactionEvent.
 
-use std::future::Future;
+use crate::module::ocpp::server::v2_common::define_csms_handler;
 
-use serde_json::{Value, json};
-
-use ferrowl_ocpp::csms::{ConnectionId, CsmsActionHandler};
-use ferrowl_ocpp::{Action201, CallError, CallErrorCode, Response201, V2_0_1, Version};
-
-use crate::module::ocpp::client::backend::rfc3339_now;
-use crate::module::ocpp::scope::Scope;
-use crate::module::ocpp::server::backend::{
-    EventTx, RfidLists, ServerEvent, cs_authorized, scope_authorized,
-};
-
-/// CSMS inbound handler for OCPP 2.0.1.
-pub struct CsmsHandler201 {
-    tx: EventTx,
-    rfids: RfidLists,
-}
-
-impl CsmsHandler201 {
-    pub fn new(tx: EventTx, rfids: RfidLists) -> Self {
-        Self { tx, rfids }
-    }
-
-    /// `"Accepted"` / `"Invalid"` for a CS-wide check (Authorize carries no EVSE).
-    fn cs_status(&self, id_token: &str) -> &'static str {
-        if cs_authorized(&self.rfids, id_token) {
-            "Accepted"
-        } else {
-            "Invalid"
-        }
-    }
-
-    /// `"Accepted"` / `"Invalid"` for a tag on a specific EVSE (its list ∪ the CS list).
-    fn evse_status(&self, evse_id: i64, id_token: &str) -> &'static str {
-        if scope_authorized(&self.rfids, Scope::evse(evse_id, None), id_token) {
-            "Accepted"
-        } else {
-            "Invalid"
-        }
-    }
-
-    fn respond(
-        &self,
-        name: &str,
-        action: &Action201,
-        request: &Value,
-    ) -> Result<Response201, CallError> {
-        let crafted: Option<Value> = match name {
-            "BootNotification" => Some(json!({
-                "currentTime": rfc3339_now(),
-                "interval": 300,
-                "status": "Accepted",
-            })),
-            "Heartbeat" => Some(json!({ "currentTime": rfc3339_now() })),
-            // Authorize carries no EVSE, so it is checked against the CS list unioned with every
-            // connector list.
-            "Authorize" => {
-                let tag = request["idToken"]["idToken"].as_str().unwrap_or_default();
-                Some(json!({ "idTokenInfo": { "status": self.cs_status(tag) } }))
-            }
-            // A TransactionEvent carrying an idToken (a transaction start) names an EVSE, so it is
-            // gated by that EVSE's list ∪ the CS list; the reply echoes the decision.
-            "TransactionEvent" if request["idToken"].is_object() => {
-                let tag = request["idToken"]["idToken"].as_str().unwrap_or_default();
-                let evse = request["evse"]["id"].as_i64().unwrap_or_default();
-                Some(json!({ "idTokenInfo": { "status": self.evse_status(evse, tag) } }))
-            }
-            _ => None,
-        };
-        match crafted {
-            Some(payload) => V2_0_1::decode_result(action, payload)
-                .map_err(|e| CallError::new(CallErrorCode::InternalError, e.to_string())),
-            None => V2_0_1::default_response(name).ok_or_else(|| {
-                CallError::new(
-                    CallErrorCode::NotImplemented,
-                    "action not handled by the CSMS",
-                )
-            }),
-        }
-    }
-}
-
-impl CsmsActionHandler<V2_0_1> for CsmsHandler201 {
-    fn handle_call(
-        &self,
-        conn: ConnectionId,
-        action: Action201,
-    ) -> impl Future<Output = Result<Response201, CallError>> + Send {
-        let name = V2_0_1::action_name(&action).to_string();
-        let request = V2_0_1::encode_action(&action).unwrap_or(Value::Null);
-        let result = self.respond(&name, &action, &request);
-        let response = match &result {
-            Ok(resp) => V2_0_1::encode_response(resp).unwrap_or(Value::Null),
-            Err(_) => Value::Null,
-        };
-        let _ = self.tx.send(ServerEvent::Inbound {
-            conn,
-            name,
-            request,
-            response,
-        });
-        async move { result }
-    }
-
-    fn on_connected(&self, conn: ConnectionId) -> impl Future<Output = ()> + Send {
-        let _ = self.tx.send(ServerEvent::Connected { conn });
-        async {}
-    }
-
-    fn on_disconnected(&self, conn: ConnectionId) -> impl Future<Output = ()> + Send {
-        let _ = self.tx.send(ServerEvent::Disconnected { conn });
-        async {}
-    }
-}
+define_csms_handler!(V2_0_1, Action201, Response201);
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::module::ocpp::scope::Scope;
     use crate::module::ocpp::server::backend::{RfidLists, RfidStore};
     use ferrowl_ocpp::csms::ConnectionId;
+    use ferrowl_ocpp::{Action201 as Action, Response201 as Response, V2_0_1 as Ver, Version};
+    use serde_json::json;
 
     fn rfids(store: RfidStore) -> RfidLists {
         std::sync::Arc::new(std::sync::RwLock::new(store))
     }
 
-    fn authorize(id_token: &str) -> Action201 {
-        V2_0_1::decode_call(
+    fn authorize(id_token: &str) -> Action {
+        Ver::decode_call(
             "Authorize",
             json!({ "idToken": { "idToken": id_token, "type": "ISO14443" } }),
         )
         .unwrap()
     }
 
-    fn status(resp: &Response201) -> String {
-        V2_0_1::encode_response(resp).unwrap()["idTokenInfo"]["status"]
+    fn status(resp: &Response) -> String {
+        Ver::encode_response(resp).unwrap()["idTokenInfo"]["status"]
             .as_str()
             .unwrap()
             .to_string()
@@ -153,7 +42,7 @@ mod tests {
             cs: vec!["GOOD".to_string()],
             ..Default::default()
         };
-        let handler = CsmsHandler201::new(tx, rfids(store));
+        let handler = CsmsHandler::new(tx, rfids(store));
         let r_good = handler
             .handle_call(ConnectionId(1), authorize("GOOD"))
             .await
@@ -169,7 +58,7 @@ mod tests {
     #[tokio::test]
     async fn ut_empty_rfid_list_accepts_all() {
         let (tx, _rx) = tokio::sync::mpsc::unbounded_channel();
-        let handler = CsmsHandler201::new(tx, rfids(RfidStore::default()));
+        let handler = CsmsHandler::new(tx, rfids(RfidStore::default()));
         let resp = handler
             .handle_call(ConnectionId(1), authorize("ANYTHING"))
             .await
@@ -186,9 +75,9 @@ mod tests {
             ..Default::default()
         };
         store.add(Scope::evse(1, None), "ETAG".to_string());
-        let handler = CsmsHandler201::new(tx, rfids(store));
+        let handler = CsmsHandler::new(tx, rfids(store));
         let started = |evse: i64, tag: &str| {
-            V2_0_1::decode_call(
+            Ver::decode_call(
                 "TransactionEvent",
                 json!({
                     "eventType": "Started",
